+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 1.0.3
+
+* Login default event
+* Register default event
+* StartView InApp Message
+
+## 1.0.2
+
+* Update User Profile
+
 ## 1.0.1
 
 * Event Tracking